Combine TypeGeneralizer and Generalizer
